--- Resources/Locale/en-US/_NF/mail/mail.ftl
@@ -1,10 +1,0 @@
-mail-large-item-name-unaddressed = package
-mail-large-item-name-addressed = package ({$recipient})
-mail-large-desc-far = A large package.
-# DelaV - Removed "Last known location"
-mail-large-desc-close = A large package addressed to {CAPITALIZE($name)}, {$job}.
-
-### Frontier: mailtestbulk
-command-mailtestbulk = Sends one of each type of parcel to a given mail teleporter.  Implicitly calls mailnow.
-command-mailtestbulk-help = Usage: {$command} <teleporter_id>
-command-mailtestbulk-success = Success! All mail teleporters will be delivering another round of mail soon.
