+++ Resources/Prototypes/DeltaV/Entities/Objects/Devices/cartridges.yml
@@ -57,5 +57,5 @@
   - type: Cartridge
     programName: mail-metrics-program-name
     icon:
-      sprite: Nyanotrasen/Objects/Specific/Mail/mail.rsi
+      sprite: Objects/Specific/Mail/mail.rsi
       state: icon
