+++ Resources/Prototypes/Entities/Objects/Specific/Mail/Items/paper.yml
