+++ Resources/Prototypes/Nyanotrasen/Entities/Objects/Specific/Mail/base_mail.yml
@@ -11,7 +11,7 @@
   - type: AccessReader
   - type: Sprite
     scale: 0.7, 0.7 # Frontier
-    sprite: Nyanotrasen/Objects/Specific/Mail/mail.rsi
+    sprite: Objects/Specific/Mail/mail.rsi
     layers:
     - state: icon
       map: ["enum.MailVisualLayers.Icon"]
