+++ Resources/Prototypes/Recipes/Lathes/misc.yml
@@ -207,3 +207,19 @@
   materials:
     Steel: 400
     Glass: 200
+
+- type: latheRecipe
+  id: ClothingShoesBootsMagAdv
+  result: ClothingShoesBootsMagAdv
+  completetime: 12
+  materials:
+    Silver: 1000
+    Gold: 500
+
+- type: latheRecipe
+  id: MailCapsule
+  result: MailCapsulePrimed
+  completetime: 1
+  materials:
+    Glass: 100
+    Plastic: 100
