--- Resources/Prototypes/_NF/Entities/Objects/Weapons/Guns/Launchers/launchers.yml
@@ -1,24 +1,0 @@
-- type: entity
-  name: mail RPDS
-  parent: WeaponLauncherChinaLake
-  id: WeaponMailLake
-  description: Rap(b?)id Parcel Delivery System
-  components:
-  - type: Sprite
-    sprite: _NF/Objects/Weapons/Guns/Launchers/mail.rsi
-    layers:
-      - state: icon
-        map: ["enum.GunVisualLayers.Base"]
-  - type: Clothing
-    sprite: _NF/Objects/Weapons/Guns/Launchers/mail.rsi
-    quickEquip: false
-    slots:
-    - Back
-    - Belt
-    - suitStorage
-  - type: BallisticAmmoProvider
-    proto: null
-    whitelist:
-      tags:
-        - MailCapsule
-    capacity: 4
