--- Resources/Prototypes/_NF/Recipes/Lathes/misc.yml
@@ -1,15 +1,0 @@
-- type: latheRecipe
-  id: ClothingShoesBootsMagAdv
-  result: ClothingShoesBootsMagAdv 
-  completetime: 12
-  materials:
-    Silver: 1000
-    Gold: 500
-
-- type: latheRecipe
-  id: MailCapsule
-  result: MailCapsulePrimed
-  completetime: 1
-  materials:
-    Glass: 100
-    Plastic: 100
