+++ Resources/Textures/Objects/Misc/mail_capsule.rsi/icon-empty.png
