+++ Resources/Textures/Objects/Misc/mail_capsule.rsi/icon-food.png
