+++ Resources/Textures/Objects/Misc/mail_capsule.rsi/icon-mail.png
