+++ Resources/Textures/Objects/Misc/mail_capsule.rsi/meta.json
