+++ Resources/Textures/Objects/Misc/mail_capsule.rsi/spent.png
