+++ Resources/Textures/Objects/Specific/Mail/mail.rsi/icon.png
