+++ Resources/Textures/Objects/Specific/Mail/mail.rsi/meta.json
