+++ Resources/Textures/Objects/Specific/Mail/mail.rsi/priority.png
