+++ Resources/Textures/Objects/Specific/Mail/mail.rsi/priority_inactive.png
