+++ Resources/Textures/Objects/Specific/Mail/mail.rsi/trash.png
