+++ Resources/Textures/Objects/Specific/Mail/mail_large.rsi/inhand-left.png
