+++ Resources/Textures/Objects/Specific/Mail/mail_large.rsi/meta.json
