+++ Resources/Textures/Objects/Specific/Mail/mail_large.rsi/priority.png
