+++ Resources/Textures/Objects/Specific/Mail/mail_large.rsi/priority_inactive.png
