+++ Resources/Textures/Objects/Specific/Mail/mail_large.rsi/trash.png
