+++ Resources/Textures/Objects/Weapons/Guns/Launchers/mail.rsi/bolt-open.png
