+++ Resources/Textures/Objects/Weapons/Guns/Launchers/mail.rsi/equipped-BELT.png
