+++ Resources/Textures/Objects/Weapons/Guns/Launchers/mail.rsi/icon.png
