+++ Resources/Textures/Objects/Weapons/Guns/Launchers/mail.rsi/inhand-left.png
