+++ Resources/Textures/Objects/Weapons/Guns/Launchers/mail.rsi/inhand-right.png
