+++ Resources/Textures/Objects/Weapons/Guns/Launchers/mail.rsi/meta.json
